Antieduardo

+++ database/migrations/2024_12_18_213358_create_project_quote_valuations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('days');
             $table->string('unit');
-            $table->integer('metrado');
+            $table->float('metrado');
             $table->float('unit_value');
             $table->text('description');
             $table->foreignId('project_quote_id')->constrained('project_quotes')->onDelete('cascade');
